src/grafana_dashboard.py
from grafana_api.grafana_face import GrafanaFace

class GrafanaDashboard:
    def __init__(self, api_key=None, host='localhost:3000'):
        self.grafana = GrafanaFace(
            auth=api_key,
            host=host
        )

    def create_llm_panels(self):
        dashboard = {
            'dashboard': {
                'id': None,
                'title': 'Environmental Analysis Dashboard',
                'panels': [
                    self._create_current_analysis_panel(),
                    self._create_trend_analysis_panel(),
                    self._create_recommendations_panel()
                ]
            },
            'overwrite': True
        }
        self.grafana.dashboard.update_dashboard(dashboard)

    def _create_current_analysis_panel(self):
        return {
            'title': 'Current Environmental Analysis',
            'type': 'text',
            'gridPos': {'h': 8, 'w': 12, 'x': 0, 'y': 0},
            'options': {
                'content': '${current_analysis}',
                'mode': 'markdown'
            },
            'targets': [{
                'refId': 'A',
                'datasource': 'Neo4j',
                'query': '''
                MATCH (a:Analysis)
                ORDER BY a.timestamp DESC
                LIMIT 1
                RETURN a.content as current_analysis
                '''
            }]
        }

    def update_trend_analysis(self, analysis):
        # Update the trend analysis panel content
        # Implementation depends on your Grafana setup
        pass
