src/main_enhanced.py
import asyncio
from bme680_sensor import BME680Sensor
from tensorrt_analyzer import SensorAnalyzer
from neo4j_enhanced import Neo4jEnhanced
from grafana_dashboard import GrafanaDashboard

class EnhancedMonitor:
    def __init__(self, neo4j_uri, neo4j_user, neo4j_password, model_path):
        self.sensor = BME680Sensor()
        self.analyzer = SensorAnalyzer(model_path)
        self.neo4j = Neo4jEnhanced(neo4j_uri, neo4j_user, neo4j_password)
        self.grafana = GrafanaDashboard()

    async def monitor_cycle(self):
        # Get sensor readings
        readings = self.sensor.get_readings()

        # Get real-time analysis
        analysis = self.analyzer.analyze_current_readings(readings)

        # Store data with analysis
        self.neo4j.store_reading_with_analysis(readings, analysis)

        # Get historical analysis every hour
        if self._should_analyze_trends():
            history = self.neo4j.get_historical_readings()
            trend_analysis = self.analyzer.analyze_trends(history)
            self.grafana.update_trend_analysis(trend_analysis)

    def _should_analyze_trends(self):
        # Add logic to determine if it's time for trend analysis
        return True  # Implement actual timing logic

async def main():
    monitor = EnhancedMonitor(
        neo4j_uri="bolt://localhost:7687",
        neo4j_user="neo4j",
        neo4j_password="password",
        model_path="/path/to/tensorrt_model"
    )

    while True:
        await monitor.monitor_cycle()
        await asyncio.sleep(300)  # 5-minute intervals

if __name__ == "__main__":
    asyncio.run(main())
