requirements-enhanced.txt
tensorrt-llm==0.5.0
torch>=2.0.0
grafana-api-client>=2.0.0
neo4j>=5.0.0
bme680>=1.0.5
asyncio>=3.4.3
